tic-tac-toe component created and app conetent moved there. Now app will be for binding together menu component and tic tac toe component
