+++ src/app/app.component.html
@@ -1,46 +1,2 @@
-<div class="application-wrapper center">
-  <div class="board">
-    <ng-container *ngIf = "boardHandler.winningFigure != ''">
-        <ng-container *ngTemplateOutlet="winnerFound"></ng-container>
-    </ng-container>
-    <div class="rows" *ngFor = "let row of rowIds" id='{{row}}'>
-      <div [attr.data-part-of-winning-plot] = 'boardHandler.checkIfIsPartOfWinningPlot(row, col)' 
-           [attr.data-is-occupied] = 'boardHandler.checkIfCellIsOccupied(row, col)'
-            class="cell center {{row==1?'first-row':''}}
-            {{row==rowIds.length?'last-row':''}} 
-            {{col==1?'first-column':''}} 
-            {{col==colIds.length?'last-column':''}}"
-            (click) = "boardHandler.fireOnclickHandlerAtRowColumn(row, col)"
-            *ngFor = "let col of colIds" id="{{row}} {{col}}">
-            <ng-container [ngSwitch]="boardHandler.getFigureAtRowColumn(row, col)">
-                <ng-container *ngSwitchCase="'Circle'"><ng-container *ngTemplateOutlet="circle"></ng-container></ng-container>
-                <ng-container *ngSwitchCase="'Cross'"><ng-container *ngTemplateOutlet="cross"></ng-container></ng-container>
-                <ng-container *ngSwitchDefault><ng-container *ngTemplateOutlet="none"></ng-container></ng-container>
-            </ng-container>
-            <!-- <div *ngIf="boardHandler.getFigureAtRowColumn(row, col)=='circle'; then circle else cross"></div> -->
-      </div>
-    </div>
-  </div>
-</div>
 
-<ng-template #cross>
-    <div class="wrapper center">
-        <div class="right-cross"></div>
-        <div class="left-cross"></div>
-    </div>
-</ng-template>
-<ng-template #circle>
-    <div class="wrapper center">
-        <div class="circle"></div>
-    </div>
-</ng-template>
-<ng-template #none></ng-template>
-<ng-template #winnerFound>
-    <div class="board-cover center">
-        <span class="center">
-                <span *ngIf = "boardHandler.winningFigure == 'Circle'" class="winner-symbol"> &#9675;</span> 
-                <span *ngIf = "boardHandler.winningFigure == 'Cross'"  class="winner-symbol"> &times;</span>
-                <span class = "winner-caption">wins.</span>
-        </span>    
-    </div>
-</ng-template>
+<tic-tac-toe></tic-tac-toe>
