+++ src/app/app.component.ts
@@ -1,6 +1,6 @@
 import { Component, Injectable, TemplateRef } from '@angular/core';
 import { bindCallback } from 'rxjs';
-import { BoardHandlerServiceService } from './board-handler-service.service'
+// import { BoardHandlerServiceService } from './board-handler-service.service'
 
 
 @Component({
@@ -14,18 +14,10 @@ export class AppComponent {
   nrOfRows = this.boardSize;  // for simplisity and pleability let board be a square
   nrOfColumns = this.boardSize;
   nrOfFiguresInRowToWinn = 3;
-  rowIds = this.createArrayOfNElements(this.boardSize);
-  colIds = this.createArrayOfNElements(this.boardSize);
 
-  constructor(public boardHandler: BoardHandlerServiceService){
-    this.boardHandler = boardHandler;
-    this.boardHandler.parametrize(this.boardSize, this.nrOfFiguresInRowToWinn);
+  constructor(){
   }
 
-  createArrayOfNElements(n:number){
-    let output = [];
-    for(let i = 1; i < n + 1; i++){ output.push(i) }
-    return output;
-  }
+
 }
