+++ src/app/app.module.ts
@@ -3,12 +3,12 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { CrossCircleComponent } from './cross-circle/cross-circle.component';
+import { TicTacToeComponent } from './tic-tac-toe/tic-tac-toe.component';
 
 @NgModule({
   declarations: [
     AppComponent,
-    CrossCircleComponent
+    TicTacToeComponent,
   ],
   imports: [
     BrowserModule,
