--- src/app/cross-circle-inserter.service.ts
@@ -1,17 +1,0 @@
-import { 
-  Injectable,
-  ComponentFactoryResolver,
-  Inject,
-  ReflectiveInjector
-} from '@angular/core';
-import { CrossCircleComponent } from './cross-circle/cross-circle.component'
-
-@Injectable({
-  providedIn: 'root'
-})
-export class CrossCircleInserterService {
-
-  constructor(private componentFactoryResolver: ComponentFactoryResolver) {}
-
-
-}
