--- src/app/cross-circle/cross-circle.component.html
@@ -1,16 +1,0 @@
-
-<div class="cross-circle">
-    <div *ngIf="figure=='circle'; then circle else cross"></div>
-</div>
-
-<ng-template #cross>
-    <div class="wrapper">
-        <div class="right-cross"></div>
-        <div class="left-cross"></div>
-    </div>
-</ng-template>
-<ng-template #circle>
-    <div class="wrapper">
-        <div class="circle"></div>
-    </div>
-</ng-template>
