+++ src/app/tic-tac-toe/tic-tac-toe.component.css
@@ -1,0 +1,128 @@
+*{
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+.application-wrapper{
+    position: relative;
+    width: 100vw;
+    height: 100vh;
+}
+.center{
+    display: flex;
+    align-content: center;
+    align-items: center;
+    justify-content: center;
+}
+.board{
+    --cell-size: 5rem;
+    --board-width: 3;
+    --board-height: 3;
+}
+.rows{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+}
+.board{
+    display: flex;
+    flex-direction: column;
+    position: relative;
+
+}
+.cell{
+    position: relative;
+    border: solid medium black;
+    width: var(--cell-size);
+    height: var(--cell-size);
+    transition: 0.2s;
+    
+}
+.first-row {
+    border-top: none;
+}
+.last-row {
+    border-bottom: none;
+}
+.first-column {
+    border-left: none;
+}
+.last-column {
+    border-right: none;
+}
+[data-is-occupied=false]:hover{
+    background-color: rgba(0, 0, 0, 0.2);
+    transition: 0.3s;
+    cursor: pointer;
+}
+[data-is-occupied=false]:active{
+    background-color: rgba(0, 255, 0, 0.2);
+    transition: 0.3s;
+}
+[data-is-occupied=true]:hover{
+    cursor: not-allowed;
+}
+.cell-ocupied{
+    background-color: rgba(255, 0, 0, 0.2);
+    transition: 0.3s;
+}
+
+.wrapper{
+    position: relative;
+    width: var(--cell-size);
+    height: var(--cell-size);
+}
+
+.circle{
+    border-radius: 50%;
+    border: solid black thick;
+    background-color: transparent;
+    position: relative;
+    width: calc( var(--cell-size) - calc(0.25 * var(--cell-size)) );
+    height: calc( var(--cell-size) - calc(0.25 * var(--cell-size)) );
+}
+
+.right-cross{
+    position: absolute;
+    height: calc( var(--cell-size) - calc(0.25 * var(--cell-size)) );
+    width: 0px;
+    transform: rotateZ(45deg);
+    border: solid medium black;
+}
+.left-cross{
+    position: absolute;
+    height: calc( var(--cell-size) - calc(0.25 * var(--cell-size)) );
+    width: 0px;
+    transform: rotateZ(-45deg);
+    border: solid medium black;
+}
+
+[data-part-of-winning-plot=true]{
+    background-color: yellow;
+}
+
+.board-cover{
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    z-index: 10;
+}
+
+.board-cover>span{
+    transform: rotate(-35deg);
+    background-color: green;
+    padding: 1rem;
+    height: 2.5rem;
+    width: 10rem;
+    border-radius: 1.5rem;
+    font-size: 2rem;
+}
+
+.winner-symbol{
+    color: yellow;
+    font-size: 3rem;
+}
+
+.winner-caption{
+    transform: translateX(20px);
+    color: white;
+}
